pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lctl"
version = "4.1.0"
description = "Time-travel debugging for multi-agent LLM workflows"
readme = "README.md"
license = "AGPL-3.0-only"
authors = [
    { name = "LCTL Contributors" }
]
keywords = [
    "llm",
    "agents",
    "debugging",
    "observability",
    "multi-agent",
    "tracing"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Debuggers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.9"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "black>=24.0",
    "ruff>=0.5",
    "httpx>=0.27",
]
dashboard = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.30",
]
langchain = [
    "langchain>=1.2.3",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
]
crewai = [
    "crewai>=1.8.0",
]
autogen = [
    "pyautogen>=0.7.0",
]
openai-agents = [
    "openai-agents>=0.1.0",
]
pydantic-ai = [
    "pydantic-ai>=0.0.14",
]
semantic-kernel = [
    "semantic-kernel>=1.0.0",
]
all = [
    "lctl[dashboard,langchain,crewai,autogen,openai-agents,pydantic-ai,semantic-kernel]",
]

[project.scripts]
lctl = "lctl.cli.main:main"

[project.urls]
Homepage = "https://github.com/kimasplund/LLM-Context-Transfer-Language"
Documentation = "https://github.com/kimasplund/LLM-Context-Transfer-Language#readme"
Repository = "https://github.com/kimasplund/LLM-Context-Transfer-Language"
Issues = "https://github.com/kimasplund/LLM-Context-Transfer-Language/issues"

[tool.hatch.build.targets.wheel]
packages = ["lctl"]

[tool.black]
line-length = 100
target-version = ["py312", "py313", "py314"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=lctl --cov-report=term-missing"
